requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
openai
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
celery
redis
uvicorn
fastapi
openai-agents
openai-agents[litellm]
tweepy
Pillow
moviepy
firecrawl-py
crewai
crewai-tools